requirements.txt
Flask==3.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
textblob==0.17.1
python-dateutil==2.8.2
gunicorn==21.2.0